refactor: Pawn Capture 로직 추상화
