+++ src/main/java/chess/domain/piece/Pawn.java
@@ -3,12 +3,12 @@ package chess.domain.piece;
 import chess.domain.Board;
 import chess.domain.Color;
 import chess.domain.piece.strategy.MovingStrategy;
-import chess.domain.piece.strategy.pawn.BlackPawnCaptureMovingStrategy;
 import chess.domain.piece.strategy.pawn.BlackPawnDefaultMovingStrategy;
 import chess.domain.piece.strategy.pawn.BlackPawnStartingPointMovingStrategy;
-import chess.domain.piece.strategy.pawn.WhitePawnCaptureMovingStrategy;
+import chess.domain.piece.strategy.pawn.PawnCaptureMovingStrategy;
 import chess.domain.piece.strategy.pawn.WhitePawnDefaultMovingStrategy;
 import chess.domain.piece.strategy.pawn.WhitePawnStartingPointMovingStrategy;
+import chess.domain.position.Direction;
 import chess.domain.position.Position;
 import java.util.List;
 
@@ -17,13 +17,13 @@ public class Pawn extends Piece {
     private static final List<MovingStrategy> BLACK_STRATEGIES = List.of(
             new BlackPawnStartingPointMovingStrategy(),
             new BlackPawnDefaultMovingStrategy(),
-            new BlackPawnCaptureMovingStrategy()
+            new PawnCaptureMovingStrategy(List.of(Direction.BOTTOM_LEFT, Direction.BOTTOM_RIGHT))
     );
 
     private static final List<MovingStrategy> WHITE_STRATEGIES = List.of(
             new WhitePawnStartingPointMovingStrategy(),
             new WhitePawnDefaultMovingStrategy(),
-            new WhitePawnCaptureMovingStrategy()
+            new PawnCaptureMovingStrategy(List.of(Direction.TOP_LEFT, Direction.TOP_RIGHT))
     );
 
     private final List<MovingStrategy> movingStrategies;
